# Use esbuild to bundle JavaScript code