### What I did

1. Replace browserify + uglify by esbuild
2. Remove unnecessary development dependencies
3. Integrate esbuild in the Gulp build (first step to avoid significant changes)
4. esbuild is configured to be compatible with:
   * chrome58
   * firefox57
   * safari11
   * edge16
5. Introduce a entry file `src/js/site.js` (as you can see we can now use ES6+ syntax)

The build is a lot faster (esbuild is written in Go with a JavaScript API) from ~12s to ~2s.
We will probably need to perform extensive tests to ensure that there is no regression.

##### Before

```
$ gulp pack             
(...)
[15:08:18] Finished 'pack' after 12 s
```

##### After

```
$ gulp pack
(...)
[15:00:00] Finished 'pack' after 1.69 s
```

Files generated by esbuild are roughly equivalent to browserify + uglify in size.