@Mogztter 2021-02-05
I also have another prototype based on https://rollupjs.org/ (instead of esbuild) which relies on a Babel plugin to transpile the code and https://github.com/terser/terser to minify the code.