@adam-cowley 2021-02-08
I've merged #41 - Can I just close this one without merging or is there anything we need from it?