@Mogztter 2021-02-08
> I've merged #41 

Cool :+1: 

> Can I just close this one without merging or is there anything we need from it?

Yes, we can close this pull request, it was only intended to showcase benefits and drawbacks of `esbuild` compared to `rollup`.